backend/requirements.txt
# FastAPI + server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# LLM
openai>=1.50.0

# RAG (optional — only needed if using document upload)
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
unstructured>=0.15.0

# Events feed scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0

# PDF support (optional)
pypdf>=4.0.0

# Google Sheets (milestone history from Gravity Forms)
google-api-python-client>=2.100.0
google-auth>=2.25.0

# Email (notifications via Resend)
resend>=2.0.0
